pyproject.toml
[tool.poetry]
name = "soukai-issue-creator"
version = "0.1.0"
description = "総会文書 Issue 作成プログラム"
authors = ["ygkn <2000ygkn0713@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^3.9.2"
black = "^21.9b0"


[tool.black]
line-length = 79
include = 'src/\.pyi?$'

[flake8]
ignore = "E203, E266, E501, W503, F403, F401"
max-line-length = 79
max-complexity = 18
select = "B,C,E,F,W,T4,B9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
